Refactored core math types for clarity; New constants.
Renamed the positional XYZ enclosing structs into Spatial (Spatial2D, Spaitial3D).
Added a new Euler3D for rotations represented according to degrees of rotation around the 3 basis axes.
Moved the functions which only made sense in a euler based rotation class into Euler3D, and the functions which didn't make sense for a Euler rotation into Spatial3D.
Added new constants for use throughout the library.
Added a test designed to check the memory layout of the core math types such that they remain compatible.

Also fixed a couple of annoying warnings (could have just suppressed them)

+++ Source/Platform/Generic/GenericMemory.h
@@ -16,12 +16,12 @@ struct GenericMemory
 {
 	static FORCEINLINE void* Memmove(void* Destination, const void* Source, uint64 Amount)
 	{
-		return ::memmove(Destination, Source, Amount);
+		return ::memmove(Destination, Source, (size_t)Amount);
 	}
 
 	static FORCEINLINE int32 Memcmp(const void* Destination, const void* Source, uintptr Amount)
 	{
-		return ::memcmp(Destination, Source, Amount);
+		return ::memcmp(Destination, Source, (size_t)Amount);
 	}
 
 	template<typename T>
